+++ examples/use_rust/nrf52832_ble/build.rs
@@ -55,6 +55,7 @@ fn main() {
 
 fn generate_vial_config() {
     // Generated vial config file
+    println!("cargo:rerun-if-changed=vial.json");
     let out_file = Path::new(&env::var_os("OUT_DIR").unwrap()).join("config_generated.rs");
 
     let p = Path::new("vial.json");