+++ rmk/src/keyboard.rs
@@ -13,6 +13,7 @@ use crate::{
 };
 use core::cell::RefCell;
 use defmt::{debug, error, warn};
+use embassy_futures::yield_now;
 use embassy_sync::{
     blocking_mutex::raw::CriticalSectionRawMutex,
     channel::{Receiver, Sender},
@@ -48,14 +49,9 @@ pub(crate) async fn keyboard_task<
     loop {
         let _ = keyboard.scan_matrix(sender).await;
         keyboard.send_keyboard_report(sender).await;
-        // Yield once to improve the performance
-        embassy_futures::yield_now().await;
         keyboard.send_media_report(sender).await;
-        embassy_futures::yield_now().await;
         keyboard.send_mouse_report(sender).await;
-        embassy_futures::yield_now().await;
         keyboard.send_system_control_report(sender).await;
-        embassy_futures::yield_now().await;
         Timer::after_micros(100).await;
     }
 }
@@ -201,11 +197,16 @@ impl<
         sender: &mut Sender<'a, CriticalSectionRawMutex, KeyboardReportMessage, 8>,
     ) {
         if self.need_send_key_report {
-            debug!("Sending keyboard report: {=[u8]:#X}", self.report.keycodes);
+            debug!(
+                "Sending keyboard report: {=[u8]:#X}, modifier: {:b}",
+                self.report.keycodes, self.report.modifier
+            );
             sender
                 .send(KeyboardReportMessage::KeyboardReport(self.report))
                 .await;
             self.need_send_key_report = false;
+            // Yield once after sending the report to channel
+            yield_now().await;
         }
     }
 
@@ -223,6 +224,7 @@ impl<
                 .await;
             self.other_report.system_usage_id = 0;
             self.need_send_system_control_report = false;
+            yield_now().await;
         }
     }
 
@@ -240,6 +242,7 @@ impl<
                 .await;
             self.other_report.media_usage_id = 0;
             self.need_send_consumer_control_report = false;
+            yield_now().await;
         }
     }
 
@@ -271,6 +274,7 @@ impl<
                 // Release, stop report mouse report
                 self.need_send_mouse_report = false;
             }
+            yield_now().await;
         }
     }
 
@@ -439,13 +443,26 @@ impl<
         key_state: KeyState,
         sender: &mut Sender<'a, CriticalSectionRawMutex, KeyboardReportMessage, 8>,
     ) {
-        // Process modifier first
-        let (keycodes, n) = modifier.to_modifier_keycodes();
-        for kc in keycodes.iter().take(n) {
-            self.process_action_keycode(*kc, key_state, sender).await;
+        if key_state.is_pressing() {
+            // Process modifier
+            let (keycodes, n) = modifier.to_modifier_keycodes();
+            for kc in keycodes.iter().take(n) {
+                self.process_action_keycode(*kc, key_state, sender).await;
+            }
+            // Send the modifier first, then send the key
+            self.send_keyboard_report(sender).await;
+            self.process_key_action_normal(action, key_state, sender)
+                .await;
+        } else {
+            // Releasing, release the key first, then release the modifier
+            self.process_key_action_normal(action, key_state, sender)
+                .await;
+            self.send_keyboard_report(sender).await;
+            let (keycodes, n) = modifier.to_modifier_keycodes();
+            for kc in keycodes.iter().take(n) {
+                self.process_action_keycode(*kc, key_state, sender).await;
+            }
         }
-        self.process_key_action_normal(action, key_state, sender)
-            .await;
     }
 
     /// Tap action, send a key when the key is pressed, then release the key.
@@ -538,24 +555,15 @@ impl<
             self.process_action_consumer_control(key, key_state);
         } else if key.is_system() {
             self.process_action_system_control(key, key_state);
-        } else if key.is_modifier() {
+        } else if key.is_basic() {
             self.need_send_key_report = true;
-            let modifier_bit = key.as_modifier_bit();
             if key_state.pressed {
-                self.register_modifier(modifier_bit);
+                self.register_key(key);
             } else {
-                self.unregister_modifier(modifier_bit);
+                self.unregister_key(key);
             }
         } else if key.is_mouse_key() {
             self.process_action_mouse(key, key_state);
-        } else if key.is_basic() {
-            self.need_send_key_report = true;
-            // 6KRO implementation
-            if key_state.pressed {
-                self.register_keycode(key);
-            } else {
-                self.unregister_keycode(key);
-            }
         } else if key.is_macro() {
             // Process macro
             self.process_action_macro(key, key_state, sender).await;
@@ -703,31 +711,35 @@ impl<
                     match operation {
                         MacroOperation::Press(k) => {
                             self.need_send_key_report = true;
-                            self.register_keycode(k);
+                            self.register_key(k);
                         }
                         MacroOperation::Release(k) => {
                             self.need_send_key_report = true;
-                            self.unregister_keycode(k);
+                            self.unregister_key(k);
                         }
                         MacroOperation::Tap(k) => {
                             self.need_send_key_report = true;
-                            self.register_keycode(k);
+                            self.register_key(k);
                             self.send_keyboard_report(sender).await;
                             embassy_time::Timer::after_millis(2).await;
                             self.need_send_key_report = true;
-                            self.unregister_keycode(k)
+                            self.unregister_key(k)
                         }
                         MacroOperation::Text(k, is_cap) => {
                             self.need_send_key_report = true;
-                            self.register_keycode(k);
                             if is_cap {
+                                // If it's a capital letter, send shift first
                                 self.register_modifier(KeyCode::LShift.as_modifier_bit());
+                                self.send_keyboard_report(sender).await;
+                                self.need_send_key_report = true;
                             }
+                            self.register_keycode(k);
                             self.send_keyboard_report(sender).await;
-                            embassy_time::Timer::after_millis(2).await;
                             self.need_send_key_report = true;
                             self.unregister_keycode(k);
                             if is_cap {
+                                self.send_keyboard_report(sender).await;
+                                self.need_send_key_report = true;
                                 self.unregister_modifier(KeyCode::LShift.as_modifier_bit());
                             }
                         }
@@ -751,6 +763,24 @@ impl<
             } else {
                 error!("Macro not found");
             }
+        }
+    }
+
+    /// Register a key, the key can be a basic keycode or a modifier.
+    fn register_key(&mut self, key: KeyCode) {
+        if key.is_modifier() {
+            self.register_modifier(key.as_modifier_bit());
+        } else if key.is_basic() {
+            self.register_keycode(key);
+        }
+    }
+
+    /// Unregister a key, the key can be a basic keycode or a modifier.
+    fn unregister_key(&mut self, key: KeyCode) {
+        if key.is_modifier() {
+            self.unregister_modifier(key.as_modifier_bit());
+        } else if key.is_basic() {
+            self.unregister_keycode(key);
         }
     }