+++ rmk/CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Update versions of dependecies
+
+## [0.2.3] - 2024-07-25
+
+### Fixed
+
+- Fix keymap doesn't change issue
+- Fix with_modifier action doesn't trigger the key with modifier
+- Fix capital letter is not send in keyboard macro
+
+### Changed
+
+- Yield everytime after sending a keyboard report to channel
+- Update `sequential-storage` to v3.0.0
+- Update `usbd-hid` to v0.7.1
+
 ## [0.2.2] - 2024-07-12
 
 - Add keyboard macro support